requirements.txt
numpy
matplotlib
pyqt5
pyqtgraph
pyopengl
plyfile
